requirements.txt
pyshark
openai
smtplib
python-dotenv
